# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/pricing_rules:
    get:
      tags:
        - PricingRules
      summary: List PricingRules
      description: Get a paginated list of PricingRules with optional filtering and sorting
      operationId: listPricingRules
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingRuleList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PricingRules
      summary: Create PricingRule
      description: Create a new PricingRule
      operationId: createPricingRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePricingRuleRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingRule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/promo_bundle_components:
    get:
      tags:
        - PromoBundleComponents
      summary: List PromoBundleComponents
      description: Get a paginated list of PromoBundleComponents with optional filtering and sorting
      operationId: listPromoBundleComponents
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoBundleComponentList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PromoBundleComponents
      summary: Create PromoBundleComponent
      description: Create a new PromoBundleComponent
      operationId: createPromoBundleComponent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePromoBundleComponentRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoBundleComponent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PricingRule:
      type: object
      required:
        - id
        - company_id
        - title
        - priority
        - scope
        - min_order_amount
        - stackable
        - apply_on
        - min_qty
        - min_amount
        - rate_or_discount
        - currency
        - valid_from
        - coupon_required
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 140
        priority:
          type: integer
          format: int32
        scope:
          type: string
        min_order_amount:
          type: string
        min_order_qty:
          nullable: true
          type: string
        stackable:
          type: boolean
        apply_on:
          type: string
        item_id:
          nullable: true
          type: string
          format: uuid
        item_group_id:
          nullable: true
          type: string
          format: uuid
        brand_id:
          nullable: true
          type: string
          format: uuid
        customer_id:
          nullable: true
          type: string
          format: uuid
        customer_group_id:
          nullable: true
          type: string
          format: uuid
        min_qty:
          type: string
        max_qty:
          nullable: true
          type: string
        min_amount:
          type: string
        rate_or_discount:
          type: string
        rate:
          nullable: true
          type: string
        discount_percentage:
          nullable: true
          type: string
        discount_amount:
          nullable: true
          type: string
        discount_upto:
          nullable: true
          type: string
        currency:
          type: string
        valid_from:
          type: string
          format: date-time
        valid_to:
          nullable: true
          type: string
          format: date-time
        coupon_required:
          type: boolean
        status:
          type: string
        metadata:
          type: object
...
    CreatePricingRuleRequest:
      type: object
      required:
        - company_id
        - title
        - priority
        - scope
        - min_order_amount
        - stackable
        - apply_on
        - min_qty
        - min_amount
        - rate_or_discount
        - currency
        - valid_from
        - coupon_required
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 140
        priority:
          type: integer
          format: int32
        scope:
          type: string
        min_order_amount:
          type: string
        min_order_qty:
          nullable: true
          type: string
        stackable:
          type: boolean
        apply_on:
          type: string
        item_id:
          nullable: true
          type: string
          format: uuid
        item_group_id:
          nullable: true
          type: string
          format: uuid
        brand_id:
          nullable: true
          type: string
          format: uuid
        customer_id:
          nullable: true
          type: string
          format: uuid
        customer_group_id:
          nullable: true
          type: string
          format: uuid
        min_qty:
          type: string
        max_qty:
          nullable: true
          type: string
        min_amount:
          type: string
        rate_or_discount:
          type: string
        rate:
          nullable: true
          type: string
        discount_percentage:
          nullable: true
          type: string
        discount_amount:
          nullable: true
          type: string
        discount_upto:
          nullable: true
          type: string
        currency:
          type: string
        valid_from:
          type: string
          format: date-time
        valid_to:
          nullable: true
          type: string
          format: date-time
        coupon_required:
          type: boolean
        status:
          type: string
        metadata:
          type: object
...
    PricingRuleList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PricingRule'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PromoBundleComponent:
      type: object
      required:
        - id
        - company_id
        - bundle_id
        - apply_on
        - min_qty
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        bundle_id:
          type: string
          format: uuid
        apply_on:
          type: string
        item_id:
          nullable: true
          type: string
          format: uuid
        item_group_id:
          nullable: true
          type: string
          format: uuid
        brand_id:
          nullable: true
          type: string
          format: uuid
        min_qty:
          type: string
        metadata:
          type: object
...
    CreatePromoBundleComponentRequest:
      type: object
      required:
        - company_id
        - bundle_id
        - apply_on
        - min_qty
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        bundle_id:
          type: string
          format: uuid
        apply_on:
          type: string
        item_id:
          nullable: true
          type: string
          format: uuid
        item_group_id:
          nullable: true
          type: string
          format: uuid
        brand_id:
          nullable: true
          type: string
          format: uuid
        min_qty:
          type: string
        metadata:
          type: object
...
    PromoBundleComponentList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PromoBundleComponent'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []